requirements_improved.txt
# Core Telegram dependencies
telethon==1.32.1
python-dotenv==1.0.0

# Async and HTTP
aiohttp==3.9.1
aiofiles==23.2.1
asyncio-throttle==1.0.2

# Data processing
feedparser==6.0.10
pyyaml==6.0.1
jsonschema==4.20.0

# Security and validation
cryptography==41.0.7
bcrypt==4.1.2

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1 